+++ include/LedStrip.hpp
